+++ drivers/timer/include/hr_hw_timer.h
@@ -6,5 +6,6 @@
 void hr_hw_timer_init_1khz(uint32_t core_clock_hz);
 uint32_t hr_hw_timer_millis(void);
 void hr_hw_timer_delay_ms(uint32_t milliseconds);
+void hr_hw_timer_tick_isr(void);
 
 #endif /* HR_HW_TIMER_H */