+++ drivers/timer/src/hr_systick_baremetal_irq.c
@@ -1,0 +1,6 @@
+#include "hr_hw_timer.h"
+
+void SysTick_Handler(void)
+{
+    hr_hw_timer_tick_isr();
+}