+++ kernel/include/hairtos/hr_time.h
@@ -1,5 +1,8 @@
 #ifndef HR_TIME_H
 #define HR_TIME_H
+
 #include "hr_types.h"
+
 hr_tick_t hr_time_now(void);
-#endif
+
+#endif /* HR_TIME_H */